Added Header and Mission elements:
Created foundations for header and nav area.
Made nav a ul. Don't know if I structured this right with flex box and it stretches out to full screen but background image doesn't.

For mission area, need to figure how to center text in background image and how to make image increase and decrease when window is resized.